upgrade spring 2.0 parent

+++ src/main/java/com/emc/test/fibonacci/AtomicFloat.java
@@ -103,7 +103,7 @@ public class AtomicFloat extends Number {
 
     @Override
     public double doubleValue() {
-        return (double) get();
+        return get();
     }
 
     private static final int i(final float f) {
